refactor(csv-import): standardize file naming to use camelCase
Update file imports, exports, and references in tests to adopt camelCase naming convention. Adjust `tsconfig` to include `tests` folder.
